+++ apps/masterbots.ai/components/providers.tsx
@@ -8,15 +8,16 @@ import { SessionProvider } from 'next-auth/react'
 import { ThemeProvider as NextThemesProvider } from 'next-themes'
 import { ThemeProviderProps } from 'next-themes/dist/types'
 import { ModelProvider } from '@/lib/hooks/use-model'
+import { Session } from 'next-auth'
 
-export function Providers({ children, ...props }: ThemeProviderProps) {
+export function Providers({ children, session, ...props }: ThemeProviderProps & { session: Session }) {
   return (
     <NextThemesProvider {...props}>
       <ModelProvider>
       <PaymentProvider>
         <SidebarProvider>
           <TooltipProvider>
-            <SessionProvider>
+            <SessionProvider session={session}>
               <ThreadProvider>{children}</ThreadProvider>
             </SessionProvider>
           </TooltipProvider>
